refactor(token): move password reset token generation to token package
Moved GenerateResetToken from utils to token package and deleted password-reset.go. Updated controller logic to reflect the change.
