--- utils/reset_password.go
@@ -1,15 +1,0 @@
-package utils
-
-import (
-	"crypto/rand"
-	"encoding/hex"
-)
-
-func GenerateResetToken() (string, error) {
-	b := make([]byte, 32)
-	_, err := rand.Read(b)
-	if err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(b), nil
-}
